# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ID8 Operator API
  version: 0.1.0
  description: |
    MVP API contract for orchestrating prompt-to-production runs with HITL gates.
servers:
  - url: https://api.id8.local
paths:
...
  /v1/projects/{projectId}/approvals:
    post:
      summary: Approve or reject a stage
      operationId: submitApproval
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalRequest'
      responses:
        '200':
          description: Approval recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalEvent'
...
  /v1/projects/{projectId}/deploy:
    post:
      summary: Deploy the latest approved build
      operationId: deployProject
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeployRequest'
      responses:
        '202':
          description: Deploy accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentRecord'
components:
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
  schemas:
...
    ApprovalStage:
      type: string
      enum:
        - prd
        - design
        - tech_plan
        - deploy
    StitchAuthMethod:
      type: string
      enum:
        - api_key
        - oauth_access_token
    StitchAuthPayload:
      type: object
      properties:
        auth_method:
          $ref: '#/components/schemas/StitchAuthMethod'
        api_key:
          type: string
        oauth_token:
          type: string
        goog_user_project:
          type: string
...
    ApprovalRequest:
      type: object
      properties:
        stage:
          $ref: '#/components/schemas/ApprovalStage'
        decision:
          type: string
          enum:
            - approved
            - rejected
        notes:
          type: string
        stitch_auth:
          $ref: '#/components/schemas/StitchAuthPayload'
      required:
        - stage
        - decision
    DeployRequest:
      type: object
      properties:
        target:
          type: string
          default: production
        artifact_id:
          type: string
          format: uuid
...
    ApprovalEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        run_id:
          type: string
          format: uuid
        stage:
          $ref: '#/components/schemas/ApprovalStage'
        decision:
          type: string
        notes:
          type: string
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
      required:
        - id
        - project_id
        - run_id
        - stage
        - decision
        - created_by
        - created_at
    DeploymentRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        environment:
          type: string
        status:
          type: string
        url:
          type: string
        created_at:
          type: string
          format: date-time
      required:
        - id
        - project_id
        - environment
        - status
        - created_at